requirements.txt

flask>=2.3.3
flask-cors>=4.0.0
psycopg2-binary>=2.9.7
python-dotenv>=1.0.0
bcrypt>=4.0.1
PyJWT>=2.8.0
requests>=2.31.0
google-api-python-client>=2.149.0
google-auth>=2.35.0
google-auth-httplib2>=0.2.0
